+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- New _go to definition_ feature, via `vscode.DefinitionProvider` class.
+- New _go to definition_ feature, via `vscode.DefinitionProvider` class. Supports loop, functions, and regular assignment definitions.
 
 ### Fixed