+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Removed
+
+## [0.1.16] - 31-01-2025
+
+### Added
+
 - New _go to definition_ feature, via `vscode.DefinitionProvider` class. Supports loop, functions, and regular assignment definitions.
 - Provide auto-completion for defined local variables and global variables.